(NFC) Include vISA options in the .visaasm dump
When shader dump is enabled, the .asm file includes the full set of options
used by vISA to compile the shader. If vISA finalizer crashes during
compilation, however, the .asm won't get generated. Include the full options
in the .visaasm file as well to make it easier to reproduce such crashes.

Separately also avoid appending a unique subroutine label suffix if the label
already ends with that suffix. With this change the output .visaasm produced by
the offline vISA executable is now identical to the input.