Make CmpBfn pattern set subregion on source operands
CmpBfn did not call SetSourceModifiers before BFN instruction because BFN does not support source modifiers.
However, this function is also responsible for setting region of the operands, so not doing it is a bug.
Pattern will not match if there are any actual modifiers on sources, so we are safe to call it to set subregions.