Don't promote packed struct
For struct arg, promoting it to by-value could need insertValue/extractValue.
If the struct is packed, its field might be mis-aligned. If it is promoted,
insVal/extVal would be mis-aligned mov instructionst, which is not legit.
To avoid that, splitting them into multiple aligned mov (for example, byte mov)
must be used.

For now, simply disable promoting packed struct to pass-by-value.