[Autobackout][FuncReg]Revert of change: 6da2d489442871c9f99aa91a75f4be615333615d  Change sample sources execution mask
Before change: sample sources calculations were marked as 'subspanUse'
and noMask was applied. Exception - noMask was not applied if sample
result was a source to another sample, Vmask needed was set instead.
After change: sample sources are still marked as 'subspanUse', but they
list is also stored separately - subspanUse is a larger set than only
sample sources calculations. Additionaly, list of sample sources under
control flow is created. Vmask is still requested if sample result is
used as a souce for another sample.
Execution mask policy - if shader is executed with dispatch mask (Vmask
not requested), old behaviour is preserved - noMasks are applied. If
Vmask is requested, noMask is only applied for sample sources under
control flow - this is for covering common application missuses of
samples.