Unify method for assigning variable 'id' in liveness
Earlier implementation used different methods to assign variable 'id' in
LivenessAnalysis pass. This was because of local/global partitioning done based on 'id'. Now, we no longer expect such 'id' based partitioning. So this change refactors 'id' assignment of variables so we've a single implementation.

This is not an NFC change because 'id' assignment is done in order of
G4_Declare* appearance in kernel.Declare list. We use 'id' as a
tie-breaker in computing coloring order for variables that have same
spill cost. So changing variable 'id' assignment order can change
coloring order and subsequently alter assignments.