Generic load/store combine pass
This is for combining LLVM LoadInst and StoreInst. It works even
those loads and stores have different element sizes (current memopt
does not handle load/store with different element sizes).

This is the first submit. It has majority of boilerplace code
implemented. It has store combining, the load combining will be
added later.