Remove ConstBase and GlobalBase implicit arguments for PVC+ platforms
Disable ConstBase and GlobalBase implicit arguments on PVC+. The purpose of adding
those arguments is to enable the opportunity to promote global/const variables access
to stateful memory. But StatelessToStateful pass is generally not enabled for
most of the cases on PVC+ so there is no benefit to add the implicit arguments.
Without the implicit arguments, const/global variables access will go through
relocations.