New option to emit all kernels/functions of a compilation into a single file.
The change is in preparation for stopping vISA .isa binary file
generation. To support offline testing, add a new vISA option
-dumpcombinedcisa, which can be enabled by a new IGC
EnableVISADumpCombinedCISA regkey, to emit all kernels and functions of
a compilation into a single file, so that GenX_IR is able to read the
combined .isaasm back when needed.