[IGC vISA] Fix the bug in avoidInstDstSrcOverlap()
Second half of a source operand must not point to the same register as the first
half of destination operand in a compressed instruction. In avoidInstDstSrcOverlap(),
it uses wrong left/right bound to check if Second half of a source is overlapped with
the first half of destination.

For below example, HWConformiry treats it as no overlap without the fix.
.decl V001 v_type=G type=W num_elts=64 align=GRF
.decl V002 v_type=G type=ub num_elts=32 alias=<V001, 0>
mov (M1, 8) V001(0,30)<1> V002(0,21)<1;1,0>