IMadLegalization fix for VC: avoid global volatile value clobbering.
Mad legalization pass can move (and duplicate) certain instructions.
In case the instruction being moved is sourcing global volatile value,
we need to check if it will be clobbered in the new position and in such a
case cancel the movement.

Broader context:

This pass is one of multiple affected passes. Other passes to be fixed are:
    RegionCollapsing
    FuncBaling
    FuncGroupBaling
    Depressurizer
    ...