Reset sub_group_size when using CallWA
Stackcalls now supports SIMD32 by default, but if the EUFusion Call WA is required, we still need to default stackcall compilation to SIMD16.
However, when the Call WA is enabled and sub_group_size(32) is set, we need to also reset to sub_group_size(16) so that SIMD16 compilation will succeed.