Do ProgramScopeConstantResolution later in cogeden and add SROA+Instcombine to promote constants
If we still have any unresolved constants in the beginning of the codegen,
we have no change to promote them to immediates after they are lowered to
load/stores from constantBase.
So this change moves ProgramScopeConstantResolution after MemCpy lowering
and adds SROA+InstCombine to promote the constants if it becomes possible
after it.