Fix `AddressSpaceAliasAnalysis` pass
IGC has an optimization that allows to allocate private memory in
a global buffer if generic pointers are used in a kernel. It gives an
opportunity to avoid generating seperate branches for private and global
memory while resolving generic pointer memory operations. (Find more
details in [Generic Pointer Documentation](https://github.com/intel/intel-graphics-compiler/blob/master/documentation/igc/generic-pointers/generic-pointers.md#resolving-generic-address-space-pointer-accesses-at-runtime))

`AddressSpaceAliasAnalysis` pass was not aware of the fact that private
and global memory may actually be not disjoint when the optimization is
enabled.

The issue came up, because `GenericAddressDynamicResolution` pass
resolves loads/stores that operate on a generic pointer to loads/stores
that operate on a global pointer if `allocatePrivateAsGlobalBuffer` is
enabled (and there are no local to generic addrspacecasts). It may lead to
producing the following code:

```llvm
%a = alloca i8
%private_to_generic = addrspacecast i8* %a to i8 addrspace(4)*
%generic_to_global = addrspacecast i8 addrspace(4)* %private_to_generic to i8 addrspace(1)*  ; <-- this is generated by GenericAddressDynamicResolution
store i8 0, i8 addrspace(1)* %generic_to_global, align 1
store i8 0, i8* %a, align 1
```

Since `AddressSpaceAliasAnalysis` was not aware of such case, it could
treat `%a` and `%generic_to_global` as disjoint regions, because they
are in a different addressspaces. This change introduces an extra check
in `AddressSpaceAliasAnalysis` pass to verify if `allocatePrivateAsGlobalBuffer`
is enabled and if so, it makes it treat private and global memory as
joint regions.