(NFC) Change vISA options storage from unordered_map to a vector (#13794)
vISA options were stored internally as an unordered_map whose key is the
vISAOptions enum and value a struct containing the option's value/defaultvalue/
other metadata. Since we have a fixed number of vISA options and they are all
initialized each time we invoke the builder, there is no reason to use a map
and pay the lookup overhead. Change it to a vector instead and index directly
using vISAOption enum value.

Co-authored-by: Wei Chen <wei.y.chen@intel.com>