(NFC) Avoid unnecessary dynamic allocation for vISA option storage
VISAOptionsEntry class uses a union to store the different type of options
(bool, int32, int64, string). So there is no reason to have type-based
subclasses for it when we can directly check their type and select the
appropriate union field. This also avoids unnecessary dynamic allocation of an
option's value and default value, which are now inlined in the VISAOptionsLine
struct.