Allocate private memory on the stack for indirect group
If a subroutine is called by indirectly-called function, then it doesn't
have an access to `privateBase` implicit argument. The change introduces
treating indirect groups as groups with stackcalls in terms of handling
private memeory, meaning that `privateBase` implicit argument is no
required, since private memory is allocated on the stack.