New attributes for function groups
Two new attribute flags are added to function groups: hasCGAvailable and hasNestedCall.

hasCGAvailable: Differentiates between indirect calls where the full callgraph is available
(ie. Directly call a function in the IndirectCallGroup in the same module) and where the full
callgraph is not available (ie. Function pointer is used, or callee is in an external module).
This is useful for estimating both private and scratch memory used for stack allocation.

hasNestedCall: Indicates if there are nested stack calls in the FG. Useful for enabling CallWA.

This will also fix a bug when using FunctionCloningThreshold, where vISA reported scratch space
is under-reported due to callgraph not being available to vISA when calculating stack usage.