Don't extend RET__loc to loop header in augmentation
Loop BBs may contain a subroutine call. But loop BB list doesn't include
callee BBs. When extending live-interval to loop header, in augmentation to handle
divergent CF we find loop exit BB by checking if each successor of
current loop BB is still part of loop. Any successor BB not part of loop
BB list is considered as loop exit. But this should treat INIT BB of
callee as an exception because call BB's successor is not a loop exit at
all.

This change prevents unnecessary extension of live-ranges live-in to
INIT BB.