Marking loads and stores as nontemporal in assert function
This change adds a pass that marks loads and stores that are present in special "__devicelib_assert_fail" function, that is an assert handler for DPCPP, as nontemporal.
This translates to uncached send instructions on platforms with LSC.
This is needed to assure that stores to "assert buffer" are visible on the host when breakpoint is hit.