[Autobackout][FuncReg]Revert of change: a92ca47d70de43fc211b0f136ec6ece85071dc6b  IGC: Move vISA stack size decision into vISA completely (NFCI)
Let vISA check if the function groups have indirect or recursive
calls, and set the stack size to a conservative size when there are.
Before this change, this final decision is made in IGC.

This change includes:
- Add vISA function attribute ATTR_Recursive which is set by IGC
  to indicate the function contains recursion.
- In CISA_IR_Builder::summarizeFunctionInfo, set spillMemUsed to
  the conservative size when there is indirect or recursive call.

Note that in VC path, the final decision of vISA stack size is still
made in VC. The missing piece in VC path is the "has recursion" information
that prevents it from using the value reported from vISA directly.
Currently, VC would force conservative size whenever having stack call.
(ref: getConservativeVISAStackSize)