Support stackcall for DP emulation functions
For DP emulation, support stackcall.

As call is always enabled for DP emulation functions, key EnableSouroutineForEmulation is
no longer needed (it was added at time IGC only supports subroutine). Also replace key
ForceSubroutineForEmulation with a new key EmulationFunctionControl. This new key has
the same value as key FunctionControl, and is for DP emulation functions only.

The default for DP functions are still subroutine. It is stackcall if EmulationFunctionControl
is set to FLAG_FCALL_FORCE_STACKCALL (3)