Support vector type in layout struct
Allow layout struct to support vector type. This is due to
the following:
   store p,    <4xi32> v0
   store p+16, <4xi32> v1
Assuming p, v0, v1 are all uniform, the previous layout struct, which is
to hold the combined value for store, is

 <{i32, i32, i32, i32, i32, i32, i32, i32}>

in which no vector type is allowed. This results in many simd1 mov
instructions.

With this change, the new layout struct is:

  <{<4xi32>, <4xi32>}>

this will result into two simd(4) mov instructions, which is a quite
improvement over the previous one.

To support vector member in layout struct, its emit needs to handle
the new AOS vector emit.

This change adds all its support and let ldstcombine pass to generate
layout struct with a vector type member.