CheckInstrTypes pass not working with const expressions fix
In ShaderCodeGen.cpp UpdateInstTypeHint(CodeGenContext& ctx)
BreakConstantExpr pass wasn't being called before CheckInstrTypes, which
caused instructions with const expressions containing casts from generic
address space to go unnoticed, which in turn caused
GenericAddressDynamicResolution pass not to be added to the pass manager
when it was in fact needed. This fix adds the BreakConstantExpr pass in
the UpdateInstTypeHint function, fixing the issue.