Re-enable FunctionCloningThreshold for ZeBin (Try #5)
Introduced a new flag "DisableFunctionCloning" for controling Function Cloning. This is currently disabled by default due to regressions.
FunctionCloningThreshold flag is now only used to select the threshold instead of enabling/disabling the feature.

FunctionCloningThreshold was disabled due to regressions on ZeBin enabling.
Re-enable again after fixes for:
* Implicit Arg bug for LocalID
* Incorrect vISA scratch space calculation for indirect calls

This commit includes fix for stackfuncs called from different sub_group_size requirements, in which case we cannot convert to indirect call.
Also, we prevent conversion to indirect calls if the CallWA is potentially needed.