Extract code that sets LiveRange properties out of interference building
With incremental RA, interference building may be done only on BBs that
contain any changes. With BBs that have no change in code, the method
won't even be invoked. So any logic to discover LiveRange properties
based on instruction traversal of each BB must be moved out of
interference construction.