Change Store combining's default
Previously, LdStCombine uses vector size=8 for DW and 4 for QW.
This is to change it to 4 for DW and 2 for QW, as there are perf
regression.

Add a new key MaxStoreVectorSizeInBytes. Its default is 0, meaning
complier decides what vector size to use, which is 16 bytes. Otherwise,
use this value as the max vector size in bytes. Expect to use 32 for
vector size = 8 for dw and 4 for qw. Note that other values, too small
or over 32B, are ignored.