Implement Support for SIMD variant compilation
Indirectly-called functions that are called from kernels with different SIMD size requirements must be compiled to multiple SIMD sizes to work correctly.
This change adds support for this feature.

Indirectly-called functions are compiled to a default SIMD size based on HW platform.
If sub_group_size requirements are present, these functions are also cloned for each SIMD variant.
SIMD variant functions have mangled names to prevent clobbering the Symbol Table.
Symbols for Global Variables are only attached to the DEFAULT dummy kernel, and are not SIMD variant.
Support is also added for FunctionCloningControl, which allows direct calls converted to indirect calls to also have SIMD variants.

Note: This commit does not enable the feature (EnableSIMDVariantCompilation). It will be enabled in a separate commit.