Limit opencl-clang mutex only to Windows
It seems that opencl-clang mutex is not necessary on Linux,
since it was commited to avoid thread-safety issues that were
only observable on Windows. Additionally, the mutex causes
compilation time regressions on some Linux workloads.

This change restricts the mutex only to Windows. Eventually
there is a plan to totally remove the mutex as soon as
thread-safety issues in opencl-clang are resolved.