Improve RA optimization for scalar payload in send
For some send instruction, scalar payload such as address operand in
block load is defined as GRF aligned in vISA. This causes more RA
constraints which may result in spilling. This patch improves RA
optimization to relax the alignment restriction for the original
variable and replace local reference with temp aligned variable to
mitigate the problem.