Add padding offset to scratch space for hashing issueq
This is to resolve an issue of DG2's cache hashing issue with padding
offest. The current observation is that the perf drops when the system
requires a large scratch space size. SkipPaddingScratchSpaceSize is for
skipping the padding if the scratch space is under the limit; otherwise,
insert the padding to the begining of the kernel.