[Autobackout][FuncReg]Revert of change: a8fedcda9086d2af1182426f445019567ffc967a  Keep correct alignment of allocas in private mem on stack
If the private memory is on the stack there may be a situation where
some extra data is placed at the beginning of stack frame (e.g. prev FP).
In that case, allocas' alignment may not be satisfied. To prevent this,
a padding is added between that extra data and the private memory.