Abort when Scratch Space size exceeds HW limit on last retry attempt only
When Scratch Space size exceeds HW limit, we also need to check if the current compilation is the final retry attempt.
On retry, we may end up reducing spills and consequently SS usage such that the HW limit is not hit. Only abort compilation if the final try still has SS exceeding the HW limit.