[IGC vISA] Fix byte swizzle restriction on src0 in HWConformity
When fix byte swizzle restrictions, we split SIMD2 instruction to fix the illegal
operands to generate better codes. But if dst and src have overlap, we should avoid
splitting instruction, otherwise there will be a extra mov instructions inserted to
resolve the dst/src overlap which is not legalized.

Before fix:
   mov (2)   V42(0,14)<1>:w  V42(0,14)<2;1,0>:w
   =>
   mov (2)   TV(0,0)<1>      V42(0,14)<2;1,0>:w
   mov (1)   V42(0,14)<1>:w  TV(0,0)<0;1,0>:w
   mov (1)   V42(0,15)<1>:w  TV(0,1)<0;1,0>:w

After fix:
   mov (2)   V42(0,14)<1>:w  V42(0,14)<2;1,0>:w
   =>
   mov (2)   TV(0,28)<2>:w   V42(0,14)<2;1,0>:w
   mov (2)   V42(0,14)<1>:w  TV(0,28)<2;1,0>:w