Fixed vector member of layout struct
As <3 x i32> has alloc size of 16 bytes, it will not be packed if it is a
member of a layout struct. To avoid adding this vector into layout struct,
checking if a vector's store size is the same as its alloc size, if not,
this vector cannot be a member of layout struct.

The problem is shwon in this case :
   struct <{ <3 x float>, i32 }>
its size is 5 x 4 = 20 bytes, not 16 bytes. But layout struct requires
its size to be 16 bytes as it must be packed.