Treat spilled relocatable operands specially
When RA spills a variable that is known to always contain a compile time constant, we
optimize the fill by reloading the constant using a mov rather than
loading from memory.

However, this optimization is illegal when constant being written in to
spilled variable is a relocatable value. This change prevents spill/fill
optimization for such relocatable operands.