Fix mangling for Joint Matrix el-wise ops
During SPIRV to LLVMIR translation, Joint Matrix type in
element wise operations was mangled always to i64.
It caused conflicts when element wise operations
were used on different matrix types in the same kernel,
like printing A, B and C elements, or adding a number to
A, B and C to each element in the same kernel.
Mangled function was looking always like that:
__builtin_spirv_OpVectorExtractDynamicJointMatrixINTEL_i64_i64.
Now it is mangled like that:
__builtin_spirv_OpVectorExtractDynamicJointMatrixINTEL_packedA_8x32_i8_i64.