vISA: Update PreSched cycle estimation.
This change updates the cycle estimation to include stall after
scheduling a node. Before this change, CurrentCycle might be advanced to
the ready cycle of a not yet scheduled top node in the ready-list when
skipping hold-list. However, the top node of the ready-list might not be
the actual node to be scheduled next when adding a new scheduling
heuristic.