Stack overflow detection debug flag
This change adds StackOverflowDetection debug flag.
With this feature enabled, stack calls are instrumented with calls to detection subroutine,
that checks if the current stack pointer - stack base for current thread is not greater than
allocated stack size per thread.

Every entry point (kernel) is instrumented with an "init" function, that writes the allocated
stack size per thread and stack base per thread to a buffer. Special buffer use is needed, as the value
will need to be accessed in the stack call callees, also in "separate compilation" scenarios.

Assert buffer is reused for this functionality, as agreed with Neo runtime.
This way we don't need special handling of this feature in Neo.

When stack overflow is detected, non-zero value is written to "flag" field of AssertBufferHeader.
This results in a host-side abort in the Neo runtime.