vISA: Refactor PreRA scheduling to get rid of priority queue.
The motivation of the change is that we plan to add more heuristics
where iterating the whole ready-list/hold-list is required to pick
a scheduling candidate. So, the priority queue might not be suitable any
more with the scenarios as the heuristic would require popping nodes
from the list, picking a node, and then pushing the rest back. We adopt
the design similar to llvm machine scheduler to address this problem in
this change.