[Autobackout][FuncReg]Revert of change: e0e7ff6e2125d67f0824a1dce7aefc4f41176426  Treat spilled relocatable operands specially
When RA spills a variable that is known to always contain a compile time constant, we
optimize the fill by reloading the constant using a mov rather than
loading from memory.

However, this optimization is illegal when constant being written in to
spilled variable is a relocatable value. This change prevents spill/fill
optimization for such relocatable operands.