(NFC) Clean up spill module's constructor
Clean up SpillManagerGRF's constructor to avoid passing unnecessary arguments
(e.g., if it could be derived from other arguments). Fix a bug discovered during
refactor where we are initializing a const reference to a temporary object. We
didn't hit the bug before as the data structure is apparently never used in that
code path.

Also avoid keeping count of number of occurrences of each type of spill temp;
they are only used as debug variable names for IR dump, so we don't want to pay
compilation overhead for them.