IGC: Update to emit vISA combined .isaasm in shader dump.
Update VISABuilder::Compile() to take .isaasm file instead of .isa file.
The change is for the preparation of vISA binary deprecation, and starts
to emit the combined .isaasm in shader dump. We expect GenX_IR should be
the only user of .isa file from VISABuilder::Compile(). With the change,
it is possible for GenX_IR to replace .isa with .isaasm.