Make removeRedundMov to keep the last inst of the atomic sequence.
The last instruction of an atomic sequence does not have the atomic
attribute, and we need to cover this case if it is also a candidate to
be removed in removeRedundMov pass. This change adds doNotDelete
property that can be used to mark instructions when forming the atomic
sequence and simplify the checks for deletion in removeRedundMov.