[IGC vISA] Fix color odering issue of failSafe RA
For failSafe RA, totalGRFRegCount is subtracted by reservedRegSize in
reserveGRFSpillReg() function. In determineColorOrdering(), numColor and
availColor are initialized by totalGRFRegCount. And availColor is subtracted
by reservedRegSize again as reservedRegSize has been added into forbidden vect.
This may cause availColor overflow. So the color ordering could be totally wrong
if using the wrong availColor(a very big value) for deciding the var to be in
unconstrainedWorklist or constrainedWorklist. Fix by adding a new var named
"reserveSpillGRFCount" in GraphColor.