If statement instead of assertion fix
In ByteCodeReaderNG.cpp there is an if statement which if evaluates to
true assigns a nullptr some value. Otherwise the pointer remains NULL.
The code then goes on to call a function that dereferences the pointer.
Since the if statement is never meant to evaluate to false, an assert is
more suitable.