Zero extend the offset used for private memory handling
This is a fix for a sporadic issue reproducible on PVC. The issue is that
private memory offset is sign extended by instruction combining before
adding it to the `privateBase` pointer. When the memory offset is big
enough, meaning that it has the most significant bit enabled, it is
incorrectly extended with ones instead zeroes. It returns in producing
incorrect pointer for private memory accesses.

The incorrect sign extend instruction is inserted by Instruction Combining,
since LLVM assumes that GEP offsets may be negative.
This change adds producing zero extend instruction explicitly to avoid
sign extend instruction.