[Autobackout][FuncReg]Revert of change: ccde9c02fa6d1ed45c7b599ce1b8c035de8a038d  Fix incorrect inline location emissions
This change fixes the bug with incorrect locations for llvm.dbg.value
calls in -O2.

If location for dbg.value was not emitted in location list, we should
not try to emit it at all.
This change also deletes redundant dbg instruction setting for abstract
variable and make sure that dbg instruction for dbg variable is set
after every iteration.