Mark loads/stores for private variables
Mark loads/stores which are created for fill/spill of user private variables.
Adding this mark, because during compilation the orginal addrspace is changed
(for ex. from PRIVATE to GLOBAL) and is not visible on end stages of compilation.
This will help to identify - which load/store is related for the private variables of user.

In llvm-IR load/stores it has mark "user_addrspace_priv".
In asm dumps it has mark "address space: private;"