Address patching for zeroed global variables and i64 offsets
Fix size overflows for .data and .bss zebin sections by using
64 bit integers to calculate them.
Use 64 bits to represent offsets into inlineProgramScopeOffsets,
globalBase and constBase to prevent overflows.

Add address patching support for zeroinitialized global variables.
Previously we cannot do the zero-initializer optimization if the variable is used by another global.
This is because the offsets for the zero-init variables have not been calculated at the time of patching.
With this change, we save the patching info, and patch later after those offsets have been calculated.
This enables the optimization to defer copying zero values to the global/const buffers.