Make edgeWeightGRF faster
edgeWeightGRF() is a hot method as per VTune profile. For large
tests with subroutines, the function consumed about 8% of total compile time.

In this change, we breakup edgeWeightGRF() method so that static version in header is devoid of direct memory accesses on objects and can be easily inlined in simplification phase.

This brings ~3% reduction in total compile time for large tests.