Fixed nomaskwa
NoMaskWA had a wrong wa sequence for predicated sel instruction.
As sel's predicate is not used for calculating WrEn, previous
wa sequence sets a zero predicate on sel to avoid running sel,
which would not work. The fix is to use a temp to replace sel's
dst, and move it back to the original only if there are active
lanes.

For example,

  (W&f1.0) sel (1)  r0.6<1>:f  r0.5<0;1,0>:f  r2.1<0;1,0>:f

is changed to by nomaskwa
  (W&f1.0) sel (1)  r2.3<1>:f  r0.5<0;1,0>:f  r2.1<0;1,0>:f
  (W&f0.0.any8h) mov (1)   r0.6<1>:f  r2.3<0;1,0>:f

where f0.0 is a WA flag and r2.3:f is a reserved tmp just for
wa purpose.