Fix UserAddrSpaceMD
When igc retries under ocl path, the llvm context will be destroyed. The
metadata that belongs to the llvm context will be destroyed as well.

This is to make sure UserAddrSpaceMD will be re-created at the beginning
of retry to avoid dangling metadata ptr from being reused during retry
compilation.