Disable EnableWriteOldFPToStack by default
Disable EnableWriteOldFPToStack by default, since it is no longer needed
for stack walk debug. This change is based on dlei6g proposition. Debug
patch applied.
Before this patch frontend framepointer was saved at the beginning of
the stack call. It is not needed anymore so it is redundant
and affecting performance. Debug info generation change is also needed
to ensure proper DWARF location expression emitting.