Fix undeterministic symbols order
`SmallSet` container of pointers doesn't guarantee that the order of
iteration matches the order of insertion, therefore elements order is
undeterministic. In this case, `SmallSet` container was used for keeping
symbols that were later iterated through to print them in shader dumps,
resulting in different dumps from run to run.

This change introduces the usage of `SmallSetVector` instead of `SmallSet`
container, according to LLVM docs:
"The difference between SetVector and other sets is that the order of
iteration is guaranteed to match the order of insertion into the
SetVector. This property is really important for things like sets of
pointers. Because pointer values are non-deterministic (e.g. vary across
runs of the program on different machines), iterating over the pointers
in the set will not be in a well-defined order."