Always use `bufferOffset` for unaligned pointers
When a pointer to SVM is not aligned to 4 bytes, Neo aligns it and stores
a difference between aligned and unaligned pointer into `bufferOffset`
argument, so that IGC can correctly access the memory. Neo must align
the pointer since hardware expects surface state base address to be
aligned to 4 bytes.

This change prevents optimizing `bufferOffset` argument usage out when
IGC cannot assure that a pointer is 4 bytes aligned.