[StackOverflowDetectionMove stack overflow detection cleanup further in the compilation
When stack overflow detection calls are not needed, i.e. when there are no stack callls or VLAs in the module, we remove the code related to them. This change moves the cleanup further in the compilation process, as optimizations can remove the calls that were initially marked as stack calls (e.g. via inlinging).