Properly detect string literals used in function call chains
If a `__devicelib_assert_fail` is called from a user function that is not inlined,
then IGC is not able to track a string literal to all its uses, therefore the string
literal is placed in `.data.const` instead of `.data.const.string` section.
It results in patching GPU VA instead of string literal CPU address by Neo.