Fix VectorPreProcess handling of large loads using ldraw instruction
VectorPreProcess was not able to split large ldraw instructions when that load is the input to a store instruction, because the same LLVMIR inst was being used as the map key for both instruction types when processing the split.
This change separates the mapping of loads and stores such that such ambigity would no longer be an issue.