ConstantCoalescing - properly detect load reading vector of pointers
The current version of ConstantCoalescing pass doesn't support merging
load instructions that read global pointer from memory. The pass just bails
out if such case is detected.
The creator of the implementation forgot to also detect the case when load
instruction read a vector of pointers from memory, therefore such
cases were leading to compilation crash. This change implementes proper
detection of vector of pointers, so that the pass can bail out without
any crash. It also implements a LIT test with uses cases that triggered
the crash to confirm that the pass can consume the IR without crashing.