Remove unreasonable assert from PromoteStatelessToBindless
PromoteStatelessToBindless pass supports promotion of `LoadInst`, `StoreInst`,
`GenISA_simdBlockRead` and `GenISA_simdBlockWrite`.
At the beggining of `GetAccessInstToSrcPointerMap` function, there is a logic
that checks whether currently processed `inst` is either of supported
instruction. If not, then there is a special variable `canPromoteAccess`
that is set to `false` and hence the instruction is not promoted.
Triggering an assert every time IGC finds other than supported instruction
doesn't make sense, because the code is adjusted to expect these unsupported
instructions and bail out if it's necessary.