Emit warning in MemOpt when leading instruction misses alignment
SROA LLVM pass may sometimes set a load/store alignment to 0. It happens when
deduced alignment (based on GEP instructions) matches an alignment specified
in datalayout for a specific type. It can be problematic as MemOpt merging
logic is implemented in a way that a product of merging inherits an alignment
from the leading load/store. It results in creating memory instruction with
different type, without alignment set, therefore the information about the
correct alignment gets lost.