Limit functions promoted for FunctionCloningControl
FunctionCloningControl promotes stack functions to indirect calls to prevent cloning and decreases compilation time.
However there are certain limitations on which functions can be promoted.

For example, in the case that this change addresses, indirectly called functions cannot access "localSLM" global variables as the index into those globals must be local to the same module which initialized the SLM buffer. For this reason, a stack function that accesses SLM cannot be promoted to indirect call, and neither can any of its callers in the call chain.

This commit detects functions that cannot be promoted, and also recursively detects all callers in the call chain.