Improve performance of SynchronizationObjectCoalescing
This change is to make faster of searching through basic blocks in SynchronizationObjectCoalescing. Properties of memory access and synchronization instruction are cached in seperate containers. Moreover, there is reduced the number of used temporary containers. Additionally, each redundancy is optimized one by one so SynchronizationObjectCoalescingAnalysis is redundant and merged with SynchronizationObjectCoalescing.