[Autobackout][FuncReg]Revert of change: 8b6ece0d6c0bcb940b01432cef2eae12f18322a3  Improve performance of SynchronizationObjectCoalescing
This change is to make faster of searching through basic blocks in SynchronizationObjectCoalescing. Properties of memory access and synchronization instruction are cached in seperate containers. Moreover, there is reduced the number of used temporary containers. Additionally, each redundancy is optimized one by one so SynchronizationObjectCoalescingAnalysis is redundant and merged with SynchronizationObjectCoalescing.