Init metadata
Let new store keep the metadata of the original leading store
(Assuming all merged stores will have the same metadata as the leading
one; otherise, they might not be able to be merged).

In additional, when generatining InsertElement/extractelement, using i32
index (constant) instead of i64. This is to be consistent with preferring
i32 over i64 in general.