Fix layout struct copy
For struct type, GetType() (to get its visa type) is byte always. When doing
struct layout copy, need to use int type that is the same size as its member.

For example, a copy from uniform Src to uniform Dst, both are of the struct
type as below:
    struct.soa { struct.aos {i8, i8, i8 ,i8}, struct.aos {i8, i8, i8, i8} }
As its member is 4 bytes, the copy will be using dw as copy type:
    mov (2)  Dst:d  Src:d