Fix alignment of source used by `lsc.store` that pushes frame-pointer to the stack
The source operand of `lsc.store` must be GRF-aligned. IGC was missing
the logic that would realign the variable used for keeping the caller
frame's frame-pointer.

Without this fix, `FP_1` variable is `qword` aligned:
```
.decl FP_1 v_type=G type=uq num_elts=1 align=qword
lsc_store.ugm (M1_NM, 1)  flat[FP_2]:a64  FP_1:d64t
```
Then after register allocation, it becomes `r59.1` that is not GRF-aligned.

This fix changes the alignment of `FP_1` to be equal to GRF-width:
```
.decl FP_1 v_type=G type=uq num_elts=1 align=wordx32
```