[IGC Core] Refactor & fix MarkAsOutput functions
Refactor mark as output mechanism. The purpose of this logic is to mark
perThreadOffset and privateBase variables as output so that their living
time will be extented to the end of the function. Marking is happening
for -O0 debug and non-debug runs. We add FP variable to VISAModule. In
the next parts debug emitter decides wheter it should use FP or
basePointer address mechanism.
This change also fixes a bug which was causing PTO and private base
variables not being marked for O0 non-debug runs. This cause differences
in debug and non-debug code.