Improve memopt for load
Background:
  GEP canonicalization helps in some cases, but it also hurts in other
  cases. As there is a way to improve cases with the canonicalization off,
  but no clean way to improve for cases with canonicalization off, the plan
  is to turn it off by default.

This change is to improve cases when canonicalization is off. The key
EnableMemOptGEPCanon is temporarily used to ease the process of turning
it off.

This change will check if comparison of two GEPs can be changed to comparison
of last indices of those two GEPs.  If two GEPs have all operands identical
except the last one, which are either sext and zext, the comparison could be
changed to the cmparison of last indices. Note this special handling only
kicks in if the normal comparison fails.