symExpr handles sext/zext
Add a safty check on the number of values that can be evaluated.
If the number values to be evaluated is over that limit, it will
not be evaluated.

Add overflow check. By default, no overflow check as address calculation
would not make sense for address overflow (surface offset overflow
might make sense ?).

Also, change value Id to be uint16_t instead of uint32_t. Also add
an enum SymCastInfo, which will be used in future.