symExpr handles zext/zext
The rule is that no symbol is reinterpreted as two things. Currently,
only sext/zext can reinterpret a symbol as either signed or unsigned.
Here handling sext/zext so that a symbol is interpreted as either signed
or unsigned, but not both. Symbol evaluation keeps track if a symbol
has been sext or zext to reenforce this rule.