Fix an OOB access in local copy propagation.
In localCopyPropagation, when the def mov can't be replaced by hoisting
its uses, we need to propagate the dst type of mov to its uses. In this
case, we'll try to create a new temp var with the propagation type to
replace the def's dst and uses' srcs. We need to make sure that the temp
var is created with a big enough size so that the optimization won't
create an OOB access.