Move bindless promotion to `StatelessToStateful` pass
The logic implemented in `PromoteStatelessToBindless` pass is deficient
and cannot handle all the cases properly. For example: if a pointer is
shared between raw load and atomic, the pointer gets changed to bindless
thereby breaking the atomic which needs A64 pointer.

The logic implemented in `StatelessToStateful` pass can handle all these
cases properly, so it seems reasonable to reuse it for bindless promotion.