PromoteBools::promoteFunction() setAttributes fix
promoteFunction method when setting attributes for the newly created
function uses old function's attribute list, which in case some
attributes got promoted causes the new function to inherit unpromoted
attributes. This commit fixes it by defining a new method that iterates
over the attribute list, checking for each whether it's been promoted
and adding it to the new function.