[IGC VC] GenXVerify pass, initial.
The aim of this pass to provide with genx-specific IR checks ensuring that IR can be considered
semantically correct from the standpoint of VC BE backend. In VC BE backend different stages of compilation
require different subsets of correctness invariants to be satisfied. This aspect is accommodated by
GenXVerifyInvariantSet-based parametrization of the pass.

At this point the pass includes a subset of checks upon genx regioning. Checks to be added are marked with TODO.
Places marked with "TODO:spec review: demark a situation when a non-conforming real-world workload IRs have been found and
it still needs to be revealed whether we need to update the spec or correct the IRs/optimizations.

In the future the pass will be populated with more checks.

The pass is being run by default in Debug build, but can also be used being built in Release* builds as a standalone or an
auxiliary tool within LIT testing or while reducing testcases with llvm-reduce.